Extract `DeprecatedPaperclipAPI` from `Spree::Image::ActiveStorageAttachment`
The module has been extracted for cleaner interface and reusability.
